desafio aventureiro!

+++ CartasSuperTrunfo.c
@@ -10,7 +10,7 @@ int main() {
     char codigo1[10];
     char cidade1[50];
     int populacao1;
-    float area1;
+    int area1;
     float pib1;
     int pontosTuristicos1;
     float densidadepopulacional1;
@@ -22,7 +22,7 @@ int main() {
     char codigo2[10];
     char cidade2[50];
     int populacao2;
-    float area2;
+    int area2;
     float pib2;
     int pontosTuristicos2;
     float densidadepopulacional2;
@@ -44,7 +44,7 @@ int main() {
     scanf("%d", &populacao1);
 
     printf("Area (km²): ");
-    scanf("%f", &area1);
+    scanf("%d", &area1);
 
     printf("PIB (em bilhoes): ");
     scanf("%f", &pib1);
@@ -75,7 +75,7 @@ int main() {
     scanf("%d", &populacao2);
 
     printf("Area (km²): ");
-    scanf("%f", &area2);
+    scanf("%d", &area2);
 
     printf("PIB (em bilhoes): ");
     scanf("%f", &pib2);
@@ -94,11 +94,11 @@ int main() {
     printf("Codigo: %s\n", codigo1);
     printf("Nome da Cidade: %s\n", cidade1);
     printf("Populacao: %d habitantes\n", populacao1);
-    printf("Area: %f km²\n", area1);
+    printf("Area: %d km²\n", area1);
     printf("PIB: %f bilhoes de reais\n", pib1);
     printf("Numero de Pontos Turisticos: %d\n", pontosTuristicos1);
-    printf("densidade populacional: %f \n",densidadepopulacional1);
-    printf("pib per capita: %f \n milhoes",pibpercapita1);
+    printf("densidade populacional: %.2f \n",densidadepopulacional1);
+    printf("pib per capita: %.2f \n milhoes",pibpercapita1);
 ;
     // Exibição da Carta 2:
     printf("\n===== CARTA 2 =====\n");
@@ -106,11 +106,11 @@ int main() {
     printf("Codigo: %s\n", codigo2);
     printf("Nome da Cidade: %s\n", cidade2);
     printf("Populacao: %d habitantes\n", populacao2);
-    printf("Area: %f km²\n", area2);
+    printf("Area: %d km²\n", area2);
     printf("PIB: %f bilhoes de reais\n", pib2);
     printf("Numero de Pontos Turisticos: %d\n", pontosTuristicos2);
-    printf("densidade populaciona: %f \n",densidadepopulacional2);
-    printf("pib per capita: %f \n milhoes",pibpercapita2);
+    printf("densidade populaciona: %.2f \n",densidadepopulacional2);
+    printf("pib per capita: %.2f \n milhoes",pibpercapita2);
 
 
     return 0;
